Use list VMs with filtering when fetching VMs
Using GET VMs call can result in 403 if the VM is not found or if the credentials are incorrect. To avoid ambiguity use list VMs call with filtering based on name or IDs which will result in empty list if the VM is not found.

Also handles mismatch in providerID in case the VM is deleted and recreated with same name again. In such cases, we query crusoe APIs to fetch the VM by name.

Updates the deployment to run CCM in host network to remove dependency on a functional CNI